Update dashboard.php
add function detail_brg

+++ application/controllers/dashboard.php
@@ -33,13 +33,13 @@ class Dashboard extends CI_Controller{
         $this->load->view('keranjang');
         $this->load->view('templates/footer');
     }
-    
-        public function hapus_keranjang()
+
+    public function hapus_keranjang()
     {
         $this->cart->destroy();
         redirect('dashboard/index');
     }
-    
+
     public function pembayaran()
     {
         $this->load->view('templates/header');
@@ -47,15 +47,33 @@ class Dashboard extends CI_Controller{
         $this->load->view('pembayaran');
         $this->load->view('templates/footer');
     }
-    
+
     public function proses_pesanan()
     {
-        $this->cart->destroy();
-        $this->load->view('templates/header');
-        $this->load->view('templates/sidebar');
-        $this->load->view('proses_pesanan');
-        $this->load->view('templates/footer');
+        $is_processed = $this->model_invoice->index();
+        if($is_processed){
+            $this->cart->destroy();
+            $this->load->view('templates/header');
+            $this->load->view('templates/sidebar');
+            $this->load->view('proses_pesanan');
+            $this->load->view('templates/footer');
+        } else 
+        {
+            echo "Maaf, Pesanan Anda Gagal diproses!!";
+        }
     }
+
+    public function detail($id_brg)
+    {
+        $data['barang'] = $this->model_barang->detail_brg($id_brg);
+        $this->cart->destroy();
+            $this->load->view('templates/header');
+            $this->load->view('templates/sidebar');
+            $this->load->view('detail_barang');
+            $this->load->view('templates/footer');
+    }
+
+
 }
 
 ?>
